Add validation that hook matches a test file
