+++ test/src/test/kotlin/graphql/nadel/tests/hooks/top-level-field-error-is-inserted-with-all-information.kt
@@ -1,4 +1,4 @@
-package graphql.nadel.tests.util.fixtures.hooks
+package graphql.nadel.tests.hooks
 
 import graphql.ErrorType
 import graphql.GraphQLError
@@ -9,13 +9,13 @@ import graphql.nadel.hooks.HydrationArguments
 import graphql.nadel.hooks.ServiceExecutionHooks
 import graphql.nadel.normalized.NormalizedQueryField
 import graphql.nadel.tests.Engine
-import graphql.nadel.tests.util.fixtures.EngineTestHook
-import graphql.nadel.tests.util.fixtures.KeepHook
+import graphql.nadel.tests.EngineTestHook
+import graphql.nadel.tests.KeepHook
 import java.util.Optional
 import java.util.concurrent.CompletableFuture
 
 @KeepHook
-class `top-level-field-error-does-not-redact-other-top-level-fields` : EngineTestHook {
+class `top-level-field-error-is-inserted-with-all-information` : EngineTestHook {
     override fun makeNadel(engine: Engine, builder: Nadel.Builder): Nadel.Builder {
         return builder.serviceExecutionHooks(object : ServiceExecutionHooks {
             override fun isFieldForbidden(
@@ -24,16 +24,17 @@ class `top-level-field-error-does-not-redact-other-top-level-fields` : EngineTes
                 variables: Map<String, Any>?,
                 userSuppliedContext: Any?,
             ): CompletableFuture<Optional<GraphQLError>> {
-                val expectedError = GraphqlErrorBuilder.newError()
-                    .message("Hello world")
-                    .path(listOf("test", "hello"))
-                    .build()
-
-                return if (normalizedField.resultKey == "foo") {
-                    CompletableFuture.completedFuture(Optional.of(expectedError))
-                } else {
-                    CompletableFuture.completedFuture(Optional.empty())
-                }
+                return CompletableFuture.completedFuture(
+                    Optional.of(
+                        GraphqlErrorBuilder.newError()
+                            .message("Hello world")
+                            .path(listOf("test", "hello"))
+                            .extensions(mapOf("test" to "Hello there"))
+                            .location(SourceLocation(12, 34))
+                            .errorType(ErrorType.DataFetchingException)
+                            .build(),
+                    ),
+                )
             }
         })
     }
